Rendering System for Test Cases HO

+++ DirectXGame/VertexShader.hlsl
@@ -3,6 +3,7 @@ struct VS_INPUT
 	float4 position: POSITION;
 	float3 color: COLOR;
 	float3 color1: COLOR1;
+
 };
 
 struct VS_OUTPUT
@@ -18,7 +19,7 @@ cbuffer constant: register(b0)
 	row_major float4x4 m_world;
 	row_major float4x4 m_view;
 	row_major float4x4 m_proj;
-	unsigned int m_time;
+	float m_time;
 };
 
 
@@ -26,9 +27,11 @@ cbuffer constant: register(b0)
 VS_OUTPUT vsmain(VS_INPUT input)
 {
 	VS_OUTPUT output = (VS_OUTPUT)0;
+
 	
-//	output.position = lerp(input.position, input.position1, (float)((sin((float)(m_time / (float)1000.0f)) + 1.0f) / 2.0f));
-	
+	//output.position = lerp(input.position, input.position1, (float)((sin((float)(m_time / (float)1000.0f)) + 1.0f) / 2.0f));
+
+
 	//WORLD SPACE
 	output.position = mul(input.position, m_world);
 	//VIEW SPACE
